feat/bugs
