+++ src/components/carousel/SampleNextArrow.jsx
@@ -2,11 +2,9 @@ import scss from "./carousel.module.scss";
 
 function SampleNextArrow({ onClick }) {
   return (
-    <>
-      <div className="left" onClick={onClick}>
-        <img id="arrayLeft" className={scss.arrowLeft} src="images/skip.png" alt="imgicoins" />
-      </div>
-    </>
+    <div className="left" onClick={onClick}>
+      <img id="arrayLeft" className={scss.arrowLeft} src="images/skip.png" alt="imgicoins" />
+    </div>
   );
 }
