+++ src/components/carousel/SamplePrevArrow.jsx
@@ -2,16 +2,9 @@ import scss from "./carousel.module.scss";
 
 function SamplePrevArrow({ onClick }) {
   return (
-    <>
-      <div id="right" onClick={onClick}>
-        <img
-          id="rightArray"
-          className={scss.arrowRight}
-          src="images/previous.png"
-          alt="LikesPigS"
-        />
-      </div>
-    </>
+    <div id="right" onClick={onClick}>
+      <img id="rightArray" className={scss.arrowRight} src="images/previous.png" alt="LikesPigS" />
+    </div>
   );
 }
