+++ src/components/carousel/SliderHead.jsx
@@ -2,14 +2,14 @@ import scss from "./carousel.module.scss";
 
 function SliderHead() {
   return (
-    <>
-      <div className={scss.header}>
+    <header>
+      <div className={scss.homesForYou}>
         <h1>Homes For You</h1>
         <span className={scss.basedOnHomes}>Based on homes you recently viewed</span>
       </div>
       <br />
       <hr className={scss.hrLine} />
-    </>
+    </header>
   );
 }
