--- src/components/home/homePage.module.scss
@@ -1,6 +1,0 @@
-.home {
-  color: rgb(73, 73, 73);
-  margin-top: -80px;
-  width: 90%;
-  margin: 0 auto;
-}
